Factor out some reflection code to a reusable location
